feat: Polyfills
